Synchronize on Signaling class
Before that change, the handler methods for the WebSocket library were
synchronized on the WebSocketAdapter while other methods in Signaling
class were synchronized on the class itself.

Also, the public SaltyRTC API is also synchronized on the Signaling
class now. With this change, the WebSocket library and the public API
are synchronized on the same class which should guarantee thread-safety.